fix: welcome screen
