+++ src/components/Items.vue
@@ -14,7 +14,7 @@
         <w-flex column justify-space-between>
           <div class="item-label">{{ $t(currentItem.label) }}</div>
           <div class="h-max" v-if="(currentItem.family === ITEM_FAMILY_SPELL || currentItem.family === ITEM_FAMILY_CUSTOM)">
-            <w-textarea v-if="canInputCustomLabel" class="item-input item-custom-label input-value" v-model="currentItem.customLabel" rows="3" />
+            <w-textarea v-if="canInputCustomLabel" class="background-white-75 item-custom-label input-value" v-model="currentItem.customLabel" rows="3" />
             <span v-else-if="currentItem.customLabel" class="item-desc">{{ currentItem.customLabel }}</span>
           </div>
           <w-flex v-else-if="currentItem.desc" align-center justify-center class="w-max">
@@ -22,7 +22,7 @@
           </w-flex>
           <div v-if="showUse && currentItem.use !== undefined && currentItem.use > 6">
             <span class="xs2">
-              <w-input class="title1 item-input input-value" v-model.number="count" />
+              <w-input class="title1 background-white-75 input-value" v-model.number="count" />
             </span>
             <span class="title1 primary">
               /{{ currentItem.use }}
