--- src/components/Rules.vue
@@ -1,11 +1,0 @@
-<template>
-  <div class="scrollable h-auto">
-  </div>
-</template>
-
-<script>
-export default {
-  name: 'Rules',
-  data () { return {} }
-}
-</script>
