+++ src/main.js
@@ -51,4 +51,10 @@ new WaveUI(app, {
 })
 app.use(i18n).use(store)
 
+app.config.globalProperties.mausritter = {
+  hirelings: null,
+  sheet: null,
+  sheetToolbar: null
+}
+
 app.mount('#app')
